+++ tests/contract_testcases/solana/account_meta.sol
@@ -2,7 +2,7 @@ import "solana" as sol;
 
 contract spl {
 	function foo() public returns (bool, address) {
-		sol.AccountMeta meta = sol.AccountMeta(address(msg.sender), true, false);
+		sol.AccountMeta meta = sol.AccountMeta(address(this), true, false);
 		return (meta.is_writable, meta.pubkey);
 	}