+++ tests/solana_tests/balance.rs
@@ -9,8 +9,8 @@ fn get_balance() {
     let mut vm = build_solidity(
         r#"
         contract c {
-            function test() public view returns (uint64) {
-                return msg.sender.balance;
+            function test(address addr) public view returns (uint64) {
+                return addr.balance;
             }
         }"#,
     );
@@ -28,7 +28,7 @@ fn get_balance() {
         },
     );
 
-    let returns = vm.function("test", &[], Some(&new));
+    let returns = vm.function("test", &[BorshToken::Address(new)]);
 
     assert_eq!(
         returns,
@@ -72,7 +72,6 @@ fn send_fails() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(returns, vec![BorshToken::Bool(false)]);
@@ -115,7 +114,6 @@ fn send_succeeds() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(returns, vec![BorshToken::Bool(true)]);
@@ -163,7 +161,6 @@ fn send_overflows() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(returns, vec![BorshToken::Bool(false)]);
@@ -214,7 +211,6 @@ fn transfer_succeeds() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(vm.account_data.get_mut(&new).unwrap().lamports, 107);
@@ -260,7 +256,6 @@ fn transfer_fails_not_enough() {
                 value: BigInt::from(104u8),
             },
         ],
-        None,
     );
     std::println!("{:?}", res);
     assert!(res.is_err());
@@ -303,7 +298,6 @@ fn transfer_fails_overflow() {
                 value: BigInt::from(104u8),
             },
         ],
-        None,
     );
     assert!(res.is_err());
 }
@@ -341,7 +335,7 @@ fn fallback() {
         vm.stack[0].abi = Some(abi);
     }
 
-    vm.function("extinct", &[], None);
+    vm.function("extinct", &[]);
 
     assert_eq!(vm.logs, "fallback");
 }
@@ -383,7 +377,6 @@ fn value_overflows() {
                 value: BigInt::from(u64::MAX as u128 + 1),
             },
         ],
-        None,
     );
     assert_eq!(res.ok(), Some(4294967296));
 
@@ -396,7 +389,6 @@ fn value_overflows() {
                 value: BigInt::from(u128::MAX),
             },
         ],
-        None,
     );
     assert_eq!(res.ok(), Some(4294967296));
 
@@ -409,7 +401,6 @@ fn value_overflows() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(returns, vec![BorshToken::Bool(false)]);