+++ tests/solana_tests/create_contract.rs
@@ -61,7 +61,6 @@ fn simple_create_contract_no_seed() {
     let bar1 = vm.function(
         "test_other",
         &[BorshToken::Address(acc), BorshToken::Address(payer)],
-        None,
     );
 
     assert_eq!(vm.logs, "bar1 says: yo from bar0");
@@ -73,7 +72,6 @@ fn simple_create_contract_no_seed() {
     vm.function(
         "call_bar1_at_address",
         &[bar1[0].clone(), BorshToken::String(String::from("xywoleh"))],
-        None,
     );
 
     assert_eq!(vm.logs, "Hello xywoleh");
@@ -124,7 +122,6 @@ fn simple_create_contract() {
     let bar1 = vm.function(
         "test_other",
         &[BorshToken::Address(seed.0), BorshToken::Address(payer)],
-        None,
     );
 
     assert_eq!(vm.logs, "bar1 says: yo from bar0");
@@ -136,7 +133,6 @@ fn simple_create_contract() {
     vm.function(
         "call_bar1_at_address",
         &[bar1[0].clone(), BorshToken::String(String::from("xywoleh"))],
-        None,
     );
 
     assert_eq!(vm.logs, "Hello xywoleh");
@@ -199,7 +195,7 @@ fn create_contract_with_payer() {
 
     vm.constructor("x", &[BorshToken::Address(payer)]);
 
-    let ret = vm.function("f", &[], None);
+    let ret = vm.function("f", &[]);
 
     assert_eq!(
         ret[0],
@@ -245,7 +241,7 @@ fn missing_contract() {
 
     let missing = account_new();
 
-    let res = vm.function_must_fail("test_other", &[BorshToken::Address(missing)], None);
+    let res = vm.function_must_fail("test_other", &[BorshToken::Address(missing)]);
     assert_eq!(res, Ok(64424509440));
 }
 
@@ -287,7 +283,6 @@ fn two_contracts() {
     let _bar1 = vm.function(
         "test_other",
         &[BorshToken::Address(seed1.0), BorshToken::Address(seed2.0)],
-        None,
     );
 
     assert_eq!(vm.logs, "bar1 says: yo from bar0bar1 says: hi from bar0");
@@ -347,7 +342,7 @@ fn account_with_space() {
 
     assert_eq!(vm.account_data.get_mut(&data).unwrap().data.len(), 3 * 102);
 
-    let ret = vm.function("hello", &[], None);
+    let ret = vm.function("hello", &[]);
 
     assert_eq!(ret[0], BorshToken::Bool(true));
 }
@@ -388,7 +383,7 @@ fn account_with_seed() {
         511 + 102
     );
 
-    let ret = vm.function("hello", &[], None);
+    let ret = vm.function("hello", &[]);
 
     assert_eq!(ret[0], BorshToken::Bool(true));
 }
@@ -439,7 +434,7 @@ fn account_with_seed_bump() {
         511 + 102
     );
 
-    let ret = vm.function("hello", &[], None);
+    let ret = vm.function("hello", &[]);
 
     assert_eq!(ret[0], BorshToken::Bool(true));
 }
@@ -478,7 +473,7 @@ fn account_with_seed_bump_literals() {
         8192
     );
 
-    let ret = vm.function("hello", &[], None);
+    let ret = vm.function("hello", &[]);
 
     assert_eq!(ret[0], BorshToken::Bool(true));
 }
@@ -524,7 +519,6 @@ fn create_child() {
     vm.function(
         "create_child",
         &[BorshToken::Address(seed.0), BorshToken::Address(payer)],
-        None,
     );
 
     assert_eq!(