+++ tests/solana_tests/simple.rs
@@ -26,7 +26,7 @@ fn simple() {
 
     vm.logs.truncate(0);
 
-    vm.function("test", &[], None);
+    vm.function("test", &[]);
 
     assert_eq!(vm.logs, "Hello from function");
 }
@@ -83,7 +83,6 @@ fn parameters() {
                 value: BigInt::from(10u8),
             },
         ],
-        None,
     );
 
     assert_eq!(vm.logs, "x is 10");
@@ -102,7 +101,6 @@ fn parameters() {
                 value: BigInt::from(102u8),
             },
         ],
-        None,
     );
 
     assert_eq!(vm.logs, "y is 102");
@@ -127,7 +125,6 @@ fn returns() {
             width: 32,
             value: BigInt::from(10u8),
         }],
-        None,
     );
 
     assert_eq!(
@@ -155,7 +152,6 @@ fn returns() {
             width: 64,
             value: BigInt::from(982451653u64),
         }],
-        None,
     );
 
     assert_eq!(
@@ -203,18 +199,18 @@ fn flipper() {
         hex::decode("6fc90ec500000000000000001800000001").unwrap()
     );
 
-    let returns = vm.function("get", &[], None);
+    let returns = vm.function("get", &[]);
 
     assert_eq!(returns, vec![BorshToken::Bool(true)]);
 
-    vm.function("flip", &[], None);
+    vm.function("flip", &[]);
 
     assert_eq!(
         vm.data()[0..17].to_vec(),
         hex::decode("6fc90ec500000000000000001800000000").unwrap()
     );
 
-    let returns = vm.function("get", &[], None);
+    let returns = vm.function("get", &[]);
 
     assert_eq!(returns, vec![BorshToken::Bool(false)]);
 }
@@ -258,7 +254,7 @@ fn incrementer() {
         }],
     );
 
-    let returns = vm.function("get", &[], None);
+    let returns = vm.function("get", &[]);
 
     assert_eq!(
         returns,
@@ -274,10 +270,9 @@ fn incrementer() {
             width: 32,
             value: BigInt::from(5u8),
         }],
-        None,
     );
 
-    let returns = vm.function("get", &[], None);
+    let returns = vm.function("get", &[]);
 
     assert_eq!(
         returns,
@@ -359,7 +354,7 @@ fn dead_storage_bug() {
 
     vm.constructor("deadstorage", &[]);
 
-    let returns = vm.function("v", &[], None);
+    let returns = vm.function("v", &[]);
 
     assert_eq!(
         returns,
@@ -427,7 +422,6 @@ contract test3 {
                 width: 32,
                 value: BigInt::from(i),
             }],
-            None,
         );
 
         assert_eq!(
@@ -451,7 +445,6 @@ contract test3 {
                 },
                 BorshToken::Bool(true),
             ],
-            None,
         );
 
         assert_eq!(
@@ -479,7 +472,6 @@ contract test3 {
                 },
                 BorshToken::Bool(false),
             ],
-            None,
         );
 
         assert_eq!(
@@ -508,7 +500,6 @@ contract test3 {
                 width: 32,
                 value: BigInt::from(i),
             }],
-            None,
         );
 
         assert_eq!(